// frontend/src/api/models/AirConditionerDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AirConditionerDto = {
    id?: number;
    roomId?: number;
    roomName?: string;
    name?: string;
    model?: string;
    status?: string;
    mode?: string;
    targetTemperature?: number;
};